move annoated_handlers to be public
