+++ docs/api/annotated_handlers.md
@@ -1,0 +1,1 @@
+::: pydantic.annotated_handlers
