+++ pydantic/__init__.py
@@ -10,14 +10,9 @@ from pydantic_core.core_schema import (
 )
 
 from . import dataclasses
-from ._internal._annotated_handlers import (
-    GetCoreSchemaHandler as GetCoreSchemaHandler,
-)
-from ._internal._annotated_handlers import (
-    GetJsonSchemaHandler as GetJsonSchemaHandler,
-)
 from ._internal._generate_schema import GenerateSchema as GenerateSchema
 from ._migration import getattr_migration
+from .annotated_handlers import GetCoreSchemaHandler, GetJsonSchemaHandler
 from .config import ConfigDict, Extra
 from .deprecated.class_validators import root_validator, validator
 from .deprecated.config import BaseConfig
