+++ pydantic/functional_serializers.py
@@ -10,7 +10,8 @@ from pydantic_core import core_schema as _core_schema
 from typing_extensions import Annotated, Literal, TypeAlias
 
 from . import PydanticUndefinedAnnotation
-from ._internal import _annotated_handlers, _decorators, _internal_dataclass
+from ._internal import _decorators, _internal_dataclass
+from .annotated_handlers import GetCoreSchemaHandler
 
 
 @dataclasses.dataclass(**_internal_dataclass.slots_true, frozen=True)
@@ -28,9 +29,7 @@ class PlainSerializer:
     return_type: Any = PydanticUndefined
     when_used: Literal['always', 'unless-none', 'json', 'json-unless-none'] = 'always'
 
-    def __get_pydantic_core_schema__(
-        self, source_type: Any, handler: _annotated_handlers.GetCoreSchemaHandler
-    ) -> core_schema.CoreSchema:
+    def __get_pydantic_core_schema__(self, source_type: Any, handler: GetCoreSchemaHandler) -> core_schema.CoreSchema:
         """Gets the Pydantic core schema.
 
         Args:
@@ -73,9 +72,7 @@ class WrapSerializer:
     return_type: Any = PydanticUndefined
     when_used: Literal['always', 'unless-none', 'json', 'json-unless-none'] = 'always'
 
-    def __get_pydantic_core_schema__(
-        self, source_type: Any, handler: _annotated_handlers.GetCoreSchemaHandler
-    ) -> core_schema.CoreSchema:
+    def __get_pydantic_core_schema__(self, source_type: Any, handler: GetCoreSchemaHandler) -> core_schema.CoreSchema:
         """This method is used to get the Pydantic core schema of the class.
 
         Args:
@@ -264,7 +261,7 @@ else:
             return Annotated[item, SerializeAsAny()]
 
         def __get_pydantic_core_schema__(
-            self, source_type: Any, handler: _annotated_handlers.GetCoreSchemaHandler
+            self, source_type: Any, handler: GetCoreSchemaHandler
         ) -> core_schema.CoreSchema:
             schema = handler(source_type)
             schema_to_update = schema
